improve code coverage by doing dual-language compilation in the compile_test
This surfaced the fact that clj and js render constants like 3.0
differently ("3.0" in Clojure, "3" in Javascript). So rather than
complicate the test, I changed that constant to 3.1, which renders
the same in both cases. This is a little sly, but it does not
seem important to insist on byte-level perfect output equivalence
for cross-compilation, when the generated code performs the
computation faithfully in both cases